+++ packages/core/src/reducers/shouldSpeakIncomingActivity.js
@@ -1,7 +1,7 @@
 import { START_SPEAKING_ACTIVITY } from '../../lib/actions/startSpeakingActivity';
 import { STOP_SPEAKING_ACTIVITY } from '../../lib/actions/stopSpeakingActivity';
 
-export default function(state = false, { type }) {
+export default function shouldSpeakIncomingActivity(state = false, { type }) {
   switch (type) {
     case START_SPEAKING_ACTIVITY:
       state = true;