+++ packages/directlinespeech/src/cognitiveServicesPromiseToESPromise.js
@@ -1,0 +1,3 @@
+export default function cognitiveServicesPromiseToESPromise(promise) {
+  return new Promise((resolve, reject) => promise.on(resolve, reject));
+}